requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
numpy>=1.24.0
opencv-python-headless>=4.9.0
Pillow>=10.0.0
requests>=2.31.0
segment-anything>=1.0
